+++ package.json
@@ -12,7 +12,7 @@
         "copyfiles:dist": "copyfiles src/cli/commands/html/sc/* dist/src/cli/commands/html/sc/ --flat",
         "test": "npm run ts:build && mocha --colors --timeout=0",
         "test-open-edge": "npm run ts:build && mocha --grep DeviceManagementClient --colors --timeout=0",
-        "test-edge-app": "npm run ts:build && mocha --grep DeviceConfiguration --colors --timeout=0",
+        "test-deployment-workflow": "npm run ts:build && mocha --grep DeploymentWorkflow --colors --timeout=0",
         "sanity": "npm run ts:build && mocha --colors --grep @sanity --timeout=0",
         "s4f": "npm run ts:build && mocha --colors --grep @s4f --timeout=0",
         "test-dev": "npm run ts:build && cross-env NODE_TLS_REJECT_UNAUTHORIZED=0 mocha --reporter=list --timeout=0",
