+++ src/api/sdk/index.ts
@@ -36,12 +36,8 @@ export * from "./model/model-management";
 export * from "./model/model-models";
 export * from "./notification-v4/notification-v4";
 export * from "./notification-v4/notification-v4-models";
-export * from "./open-edge/device/device-management";
-export * from "./open-edge/device/device-models";
-export * from "./open-edge/device-status/device-status-management";
-export * from "./open-edge/device-status/device-status-models";
-export * from "./open-edge/edge-app/edge-app-models";
-export * from "./open-edge/edge-app/edge-app-management";
+export * from "./open-edge/open-edge";
+export * from "./open-edge/open-edge-models";
 export * from "./sdi/sdi-models";
 export * from "./sdi/sdi-v4";
 export * from "./signal-calculation/signal-calculation";
