+++ test/device-configuration.spec.ts
@@ -1,6 +1,6 @@
 import * as chai from "chai";
 import "url-search-params-polyfill";
-import { DeviceConfigurationModels } from "../src/api/sdk/open-edge/device-configuration/device-configuration-models";
+import { DeviceConfigurationModels } from "../src/api/sdk/open-edge/open-edge-models";
 import { MindSphereSdk } from "../src";
 import { decrypt, loadAuth, throwError } from "../src/api/utils";
 import { setupDeviceTestStructure, tearDownDeviceTestStructure } from "./test-device-setup-utils";
